[portfolio] Refactor receipt geometry helpers
