+++ portfolio/components/ui/Figures/PhotoReceiptBoundingBox.tsx
@@ -8,6 +8,11 @@ import {
   type Point as ApiPoint,
 } from "../../../types/api";
 import { useSpring, useTransition, animated } from "@react-spring/web";
+import {
+  computeHullCentroid,
+  computeReceiptBoxFromLineEdges,
+  computeEdge,
+} from "../../../utils/geometry";
 
 // Define simple point and line-segment shapes
 type Point = ApiPoint;
@@ -189,21 +194,6 @@ const computeConvexHull = (points: Point[]): Point[] => {
   }
 
   return hull;
-};
-
-/**
- * Compute the centroid of a convex hull (matching Python implementation)
- */
-const computeHullCentroid = (hull: Point[]): Point => {
-  if (hull.length === 0) return { x: 0, y: 0 };
-
-  const sumX = hull.reduce((sum, point) => sum + point.x, 0);
-  const sumY = hull.reduce((sum, point) => sum + point.y, 0);
-
-  return {
-    x: sumX / hull.length,
-    y: sumY / hull.length,
-  };
 };
 
 /**
@@ -512,323 +502,7 @@ const findBoundaryLinesWithSkew = (
 /**
  * Find the top and bottom edges of lines at the secondary axis extremes
  */
-const findLineEdgesAtSecondaryExtremes = (
-  lines: Line[],
-  hull: Point[],
-  centroid: Point,
-  avgAngle: number
-): {
-  topEdge: Point[];
-  bottomEdge: Point[];
-} => {
-  const angleRad = (avgAngle * Math.PI) / 180;
-  const secondaryAxisAngle = angleRad + Math.PI / 2;
 
-  // Find secondary axis extremes (yellow circles)
-  let minSecondary = Infinity,
-    maxSecondary = -Infinity;
-  let topExtremeX = 0,
-    bottomExtremeX = 0;
-
-  hull.forEach((point) => {
-    const relX = point.x - centroid.x;
-    const relY = point.y - centroid.y;
-    const secondaryProjection =
-      relX * Math.cos(secondaryAxisAngle) + relY * Math.sin(secondaryAxisAngle);
-
-    if (secondaryProjection < minSecondary) {
-      minSecondary = secondaryProjection;
-      bottomExtremeX = point.x;
-    }
-    if (secondaryProjection > maxSecondary) {
-      maxSecondary = secondaryProjection;
-      topExtremeX = point.x;
-    }
-  });
-
-  // Find lines near the top and bottom extremes
-  const tolerance = 0.1; // 10% tolerance for finding nearby lines
-  const topLines = lines.filter((line) => {
-    const lineCenterX =
-      (line.top_left.x +
-        line.top_right.x +
-        line.bottom_left.x +
-        line.bottom_right.x) /
-      4;
-    return Math.abs(lineCenterX - topExtremeX) < tolerance;
-  });
-
-  const bottomLines = lines.filter((line) => {
-    const lineCenterX =
-      (line.top_left.x +
-        line.top_right.x +
-        line.bottom_left.x +
-        line.bottom_right.x) /
-      4;
-    return Math.abs(lineCenterX - bottomExtremeX) < tolerance;
-  });
-
-  // Find the actual top edge (highest Y values) and bottom edge (lowest Y values)
-  let topEdgePoints: Point[] = [];
-  let bottomEdgePoints: Point[] = [];
-
-  if (topLines.length > 0) {
-    // Get the topmost edges of lines near the top extreme
-    topLines.forEach((line) => {
-      const topY = Math.max(line.top_left.y, line.top_right.y);
-      topEdgePoints.push(
-        { x: line.top_left.x, y: topY },
-        { x: line.top_right.x, y: topY }
-      );
-    });
-  }
-
-  if (bottomLines.length > 0) {
-    // Get the bottommost edges of lines near the bottom extreme
-    bottomLines.forEach((line) => {
-      const bottomY = Math.min(line.bottom_left.y, line.bottom_right.y);
-      bottomEdgePoints.push(
-        { x: line.bottom_left.x, y: bottomY },
-        { x: line.bottom_right.x, y: bottomY }
-      );
-    });
-  }
-
-  // Fallback to hull points if no lines found
-  if (topEdgePoints.length === 0) {
-    const topHullPoint = hull.find((point) => {
-      const relX = point.x - centroid.x;
-      const relY = point.y - centroid.y;
-      const projection =
-        relX * Math.cos(secondaryAxisAngle) +
-        relY * Math.sin(secondaryAxisAngle);
-      return Math.abs(projection - maxSecondary) < 0.001;
-    });
-    if (topHullPoint) topEdgePoints = [topHullPoint];
-  }
-
-  if (bottomEdgePoints.length === 0) {
-    const bottomHullPoint = hull.find((point) => {
-      const relX = point.x - centroid.x;
-      const relY = point.y - centroid.y;
-      const projection =
-        relX * Math.cos(secondaryAxisAngle) +
-        relY * Math.sin(secondaryAxisAngle);
-      return Math.abs(projection - minSecondary) < 0.001;
-    });
-    if (bottomHullPoint) bottomEdgePoints = [bottomHullPoint];
-  }
-
-  return {
-    topEdge: topEdgePoints,
-    bottomEdge: bottomEdgePoints,
-  };
-};
-
-/**
- * Compute receipt box using skewed boundaries from all four edge analyses
- */
-const computeReceiptBoxFromLineEdges = (
-  lines: Line[],
-  hull: Point[],
-  centroid: Point,
-  avgAngle: number
-): Point[] => {
-  if (lines.length === 0) return [];
-
-  // Use the existing robust edge computation instead of our experimental approach
-  const leftEdge = computeEdge(lines, "left");
-  const rightEdge = computeEdge(lines, "right");
-
-  if (!leftEdge || !rightEdge) {
-    // Fallback to simple hull-based rectangle
-    return [
-      { x: centroid.x - 0.1, y: centroid.y + 0.1 }, // top-left
-      { x: centroid.x + 0.1, y: centroid.y + 0.1 }, // top-right
-      { x: centroid.x + 0.1, y: centroid.y - 0.1 }, // bottom-right
-      { x: centroid.x - 0.1, y: centroid.y - 0.1 }, // bottom-left
-    ];
-  }
-
-  // Find actual top and bottom boundaries using secondary axis extremes (yellow circles)
-  const { topEdge, bottomEdge } = findLineEdgesAtSecondaryExtremes(
-    lines,
-    hull,
-    centroid,
-    avgAngle
-  );
-
-  // Fit lines through the top and bottom edge points (yellow circles)
-  const topEdgeLine = topEdge.length >= 2 ? theilSen(topEdge) : null;
-  const bottomEdgeLine = bottomEdge.length >= 2 ? theilSen(bottomEdge) : null;
-
-  // Get left and right edge line parameters
-  const leftSlope =
-    (leftEdge.top.x - leftEdge.bottom.x) / (leftEdge.top.y - leftEdge.bottom.y);
-  const leftIntercept = leftEdge.top.x - leftSlope * leftEdge.top.y;
-
-  const rightSlope =
-    (rightEdge.top.x - rightEdge.bottom.x) /
-    (rightEdge.top.y - rightEdge.bottom.y);
-  const rightIntercept = rightEdge.top.x - rightSlope * rightEdge.top.y;
-
-  // Calculate intersections of skewed boundary lines
-  let topLeft, topRight, bottomLeft, bottomRight;
-
-  if (topEdgeLine) {
-    // Top edge: x = topSlope * y + topIntercept
-    // Left edge: x = leftSlope * y + leftIntercept
-    // Intersection: topSlope * y + topIntercept = leftSlope * y + leftIntercept
-    const topLeftY =
-      (leftIntercept - topEdgeLine.intercept) / (topEdgeLine.slope - leftSlope);
-    const topLeftX = leftSlope * topLeftY + leftIntercept;
-    topLeft = { x: topLeftX, y: topLeftY };
-
-    // Top edge intersect with right edge
-    const topRightY =
-      (rightIntercept - topEdgeLine.intercept) /
-      (topEdgeLine.slope - rightSlope);
-    const topRightX = rightSlope * topRightY + rightIntercept;
-    topRight = { x: topRightX, y: topRightY };
-  } else {
-    // Fallback to horizontal line at average Y
-    const avgTopY =
-      topEdge.reduce((sum, p) => sum + p.y, 0) / Math.max(topEdge.length, 1);
-    topLeft = { x: leftSlope * avgTopY + leftIntercept, y: avgTopY };
-    topRight = { x: rightSlope * avgTopY + rightIntercept, y: avgTopY };
-  }
-
-  if (bottomEdgeLine) {
-    // Bottom edge intersect with left edge
-    const bottomLeftY =
-      (leftIntercept - bottomEdgeLine.intercept) /
-      (bottomEdgeLine.slope - leftSlope);
-    const bottomLeftX = leftSlope * bottomLeftY + leftIntercept;
-    bottomLeft = { x: bottomLeftX, y: bottomLeftY };
-
-    // Bottom edge intersect with right edge
-    const bottomRightY =
-      (rightIntercept - bottomEdgeLine.intercept) /
-      (bottomEdgeLine.slope - rightSlope);
-    const bottomRightX = rightSlope * bottomRightY + rightIntercept;
-    bottomRight = { x: bottomRightX, y: bottomRightY };
-  } else {
-    // Fallback to horizontal line at average Y
-    const avgBottomY =
-      bottomEdge.reduce((sum, p) => sum + p.y, 0) /
-      Math.max(bottomEdge.length, 1);
-    bottomLeft = { x: leftSlope * avgBottomY + leftIntercept, y: avgBottomY };
-    bottomRight = {
-      x: rightSlope * avgBottomY + rightIntercept,
-      y: avgBottomY,
-    };
-  }
-
-  return [
-    topLeft, // top-left
-    topRight, // top-right
-    bottomRight, // bottom-right
-    bottomLeft, // bottom-left
-  ];
-};
-
-/**
- * Robust Theil–Sen estimator for a line x = m·y + b.
- */
-const theilSen = (pts: Point[]) => {
-  if (pts.length < 2) return { slope: 0, intercept: pts[0] ? pts[0].y : 0 };
-
-  // Collect all pair‑wise slopes.
-  const slopes: number[] = [];
-  for (let i = 0; i < pts.length; i++) {
-    for (let j = i + 1; j < pts.length; j++) {
-      if (pts[i].y === pts[j].y) continue; // avoid div‑by‑zero
-      slopes.push((pts[j].x - pts[i].x) / (pts[j].y - pts[i].y));
-    }
-  }
-  if (slopes.length === 0) {
-    // All points share the same y => true horizontal line
-    return { slope: 0, intercept: pts[0].y };
-  }
-  slopes.sort((a, b) => a - b);
-  const slope = slopes[Math.floor(slopes.length / 2)];
-
-  // Median of the intercepts.
-  const intercepts = pts.map((p) => p.x - slope * p.y).sort((a, b) => a - b);
-  const intercept = intercepts[Math.floor(intercepts.length / 2)];
-
-  return { slope, intercept };
-};
-
-/**
- * Compute an extreme hull edge (left or right) by y‑binning and Theil–Sen fit.
- */
-const computeHullEdge = (
-  hull: Point[],
-  bins = 12,
-  pick: "left" | "right"
-): { top: Point; bottom: Point } | null => {
-  if (hull.length < 4) return null;
-  // Bin by normalized y and pick extreme x in each bin
-  const binPts: (Point | null)[] = Array.from({ length: bins }, () => null);
-  hull.forEach((p) => {
-    const idx = Math.min(bins - 1, Math.floor(p.y * bins));
-    const current = binPts[idx];
-    if (!current || (pick === "left" ? p.x < current.x : p.x > current.x)) {
-      binPts[idx] = p;
-    }
-  });
-  const samples = binPts.filter(Boolean) as Point[];
-  if (samples.length < 2) return null;
-  const { slope, intercept } = theilSen(samples);
-  return {
-    top: { x: slope * 1 + intercept, y: 1 },
-    bottom: { x: slope * 0 + intercept, y: 0 },
-  };
-};
-
-/**
- * Compute either the left or right page edge by y‑binning extreme x values
- * and fitting a robust line through them.
- */
-const computeEdge = (
-  lines: Line[],
-  pick: "left" | "right",
-  bins = 6
-): {
-  top: Point;
-  bottom: Point;
-} | null => {
-  // One bucket per y‑range.
-  const binPts: (Point | null)[] = Array.from({ length: bins }, () => null);
-
-  lines.forEach((l) => {
-    // Normalised y in [0,1] (0 = bottom, 1 = top).
-    const yMid = (l.top_left.y + l.bottom_left.y) / 2;
-    const x =
-      pick === "left"
-        ? Math.min(l.top_left.x, l.bottom_left.x)
-        : Math.max(l.top_right.x, l.bottom_right.x);
-
-    const idx = Math.min(bins - 1, Math.floor(yMid * bins));
-    const current = binPts[idx];
-
-    if (!current) {
-      binPts[idx] = { x, y: yMid };
-    } else if (pick === "left" ? x < current.x : x > current.x) {
-      binPts[idx] = { x, y: yMid };
-    }
-  });
-
-  const selected = binPts.filter(Boolean) as Point[];
-  if (selected.length < 2) return null;
-
-  const { slope, intercept } = theilSen(selected);
-  return {
-    top: { x: slope * 1 + intercept, y: 1 },
-    bottom: { x: slope * 0 + intercept, y: 0 },
-  };
-};
 
 /**
  * Estimate the full receipt quadrilateral from OCR word‑level lines.
